+++ src/hooks/index.js
@@ -1,153 +1,3 @@
-import { useState, useEffect, useCallback } from 'react';
-import { useSelector } from 'react-redux';
-import { axiosWithAuth, selectToken } from '../utils/api';
-
 export * from './users';
-
-export const useParticipants = id => {
-  const [participants, setParticipants] = useState([]);
-  const token = useSelector(selectToken);
-  const fetchParticipants = useCallback(() => {
-    const fetchData = async () => {
-      const {
-        data: { body }
-      } = await axiosWithAuth(token).get(`/api/events/${id}/participants`);
-      setParticipants(body);
-    };
-    fetchData();
-  }, [id, token])
-
-  useEffect(() => {
-    fetchParticipants();
-
-  }, [fetchParticipants]);
-
-  return [participants, fetchParticipants];
-}
-
-export const useEventTeam = id => {
-  const [team, setTeam] = useState([]);
-  const token = useSelector(selectToken);
-  const fetchEventTeam = useCallback(() => {
-    const fetchData = async () => {
-      const {
-        data: { body: { members } }
-      } = await axiosWithAuth(token).get(`/api/events/${id}/team`);
-      setTeam(members);
-    };
-    fetchData();
-  }, [id, token]);
-
-  useEffect(() => {
-    fetchEventTeam();
-
-  }, [fetchEventTeam])
-
-  return [team, fetchEventTeam];
-}
-
-export const useJudges = id => {
-  const team = useEventTeam(id);
-  const judges = team.filter(t => t.role_type === 'judge');
-  return judges;
-}
-
-export const useTeams = id => {
-  const [teams, setTeams] = useState([]);
-  const token = useSelector(selectToken);
-
-  const fetchTeams = useCallback(() => {
-    const fetchData = async () => {
-      const { data: { body }
-      } = await axiosWithAuth(token).get(
-        `/api/events/${id}/participant-teams`
-      );
-      setTeams(body);
-    }
-    fetchData();
-  }, [id, token]);
-
-  useEffect(() => {
-    fetchTeams();
-
-  }, [fetchTeams]);
-
-  return [teams, fetchTeams];
-}
-
-export const useSubmissions = id => {
-  const [submissions, setSubmissions] = useState([]);
-  const token = useSelector(selectToken);
-  const fetchSubmissions = useCallback(() => {
-    const fetchData = async () => {
-      const { data: { body }
-      } = await axiosWithAuth(token).get(
-        `/api/events/${id}/projects`
-      );
-      setSubmissions(body);
-    }
-    fetchData();
-  }, [id, token]);
-
-  useEffect(() => {
-    fetchSubmissions();
-
-  }, [fetchSubmissions]);
-
-  return [submissions, fetchSubmissions];
-}
-
-export const useTeammates = id => {
-  const [teammates, setTeammates] = useState([]);
-  const token = useSelector(selectToken);
-  const fetchTeammates = useCallback(() => {
-    const fetchData = async () => {
-      const {
-        data: { body }
-      } = await axiosWithAuth(token).get(
-        `/api/events/participant-teams/${id}/members`
-      );
-      setTeammates(body);
-    }
-    fetchData();
-  }, [id, token]);
-
-  useEffect(() => {
-    fetchTeammates();
-
-  }, [fetchTeammates]);
-
-  return [teammates, fetchTeammates];
-}
-
-export const useGrades = (projectId) => {
-  const [grades, setGrades] = useState([]);
-  const token = useSelector(selectToken);
-
-  const fetchGrades = useCallback(() => {
-    const getData = async () => {
-      const { data: { body } } = await axiosWithAuth(token).get(
-        `/api/events/projects/${projectId}/grading`
-      );
-      setGrades(body);
-    }
-    getData();
-  }, [projectId, token]);
-
-  useEffect(() => {
-    fetchGrades();
-  }, [fetchGrades]);
-
-  return [grades, fetchGrades];
-}
-
-export const useCreatedTeam = (eventId, userId) => {
-  const [teams, fetchTeams] = useTeams(eventId);
-  const team = teams.find(team => team.team_lead === Number(userId));
-
-  useEffect(() => {
-    fetchTeams();
-  }, [fetchTeams]);
-
-  return team;
-} 
+export * from './events';
+export * from './project';
